quantum-tools/docs/references/ai_rnd_research_stack_pro/requirements.txt
requests
httpx
beautifulsoup4
lxml
readability-lxml
pdfminer.six
pandas
python-dateutil
tldextract
PyYAML
